TIP-860: move product mass edit readers
