+++ src/Akeneo/Pim/Enrichment/Component/Product/Connector/Reader/Database/MassEdit/FilteredProductReader.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Pim\Bundle\EnrichBundle\Connector\Reader\MassEdit;
+namespace Akeneo\Pim\Enrichment\Component\Product\Connector\Reader\Database\MassEdit;
 
 use Akeneo\Channel\Component\Model\ChannelInterface;
 use Akeneo\Channel\Component\Repository\ChannelRepositoryInterface;
@@ -10,9 +10,7 @@ use Akeneo\Pim\Enrichment\Component\Product\Converter\MetricConverter;
 use Akeneo\Pim\Enrichment\Component\Product\Exception\ObjectNotFoundException;
 use Akeneo\Pim\Enrichment\Component\Product\Manager\CompletenessManager;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
-use Akeneo\Pim\Enrichment\Component\Product\Model\ProductModelInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Query\ProductQueryBuilderFactoryInterface;
-use Akeneo\Tool\Component\Batch\Item\DataInvalidItem;
 use Akeneo\Tool\Component\Batch\Item\InitializableInterface;
 use Akeneo\Tool\Component\Batch\Item\ItemReaderInterface;
 use Akeneo\Tool\Component\Batch\Model\StepExecution;
@@ -22,11 +20,11 @@ use Akeneo\Tool\Component\StorageUtils\Cursor\CursorInterface;
 /**
  * Product reader that only returns product entities and skips product models.
  *
- * @author    Samir Boulil <samir.boulil@akeneo.com>
+ * @author    Pierre Allard <pierre.allard@akeneo.com>
  * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class FilteredProductAndProductModelReader implements
+class FilteredProductReader implements
     ItemReaderInterface,
     InitializableInterface,
     StepExecutionAwareInterface
@@ -46,9 +44,6 @@ class FilteredProductAndProductModelReader implements
     /** @var bool */
     private $generateCompleteness;
 
-    /** @var bool */
-    private $readChildren;
-
     /** @var StepExecution */
     private $stepExecution;
 
@@ -64,22 +59,19 @@ class FilteredProductAndProductModelReader implements
      * @param CompletenessManager                 $completenessManager
      * @param MetricConverter                     $metricConverter
      * @param bool                                $generateCompleteness
-     * @param bool                                $readChildren
      */
     public function __construct(
         ProductQueryBuilderFactoryInterface $pqbFactory,
         ChannelRepositoryInterface $channelRepository,
         CompletenessManager $completenessManager,
         MetricConverter $metricConverter,
-        bool $generateCompleteness,
-        bool $readChildren
+        bool $generateCompleteness
     ) {
         $this->pqbFactory = $pqbFactory;
         $this->channelRepository = $channelRepository;
         $this->completenessManager = $completenessManager;
         $this->metricConverter = $metricConverter;
         $this->generateCompleteness = $generateCompleteness;
-        $this->readChildren = $readChildren;
     }
 
     /**
@@ -162,29 +154,24 @@ class FilteredProductAndProductModelReader implements
             $filters = $filters['data'];
         }
 
-        if ($this->readChildren) {
-            $filters = array_map(function ($filter) {
-                if ('id' === $filter['field']) {
-                    $filter['field'] = 'self_and_ancestor.id';
-                }
-
-                return $filter;
-            }, $filters);
-        }
-
         return array_filter($filters, function ($filter) {
             return count($filter) > 0;
         });
     }
 
     /**
-     * @param array            $filters
-     * @param ChannelInterface $channel
+     * @param array                 $filters
+     * @param ChannelInterface|null $channel
      *
      * @return CursorInterface
      */
     private function getProductsCursor(array $filters, ChannelInterface $channel = null): CursorInterface
     {
+        $filters[] = [
+            'field' => 'entity_type',
+            'operator' => '=',
+            'value' => ProductInterface::class,
+        ];
         $options = ['filters' => $filters];
 
         if (null !== $channel) {
@@ -205,37 +192,18 @@ class FilteredProductAndProductModelReader implements
     {
         $entity = null;
 
-        while ($this->productsAndProductModels->valid()) {
+        if ($this->productsAndProductModels->valid()) {
             if (!$this->firstRead) {
                 $this->productsAndProductModels->next();
             }
 
-            $this->firstRead = false;
             $entity = $this->productsAndProductModels->current();
             if (false === $entity) {
                 return null;
             }
-
-            if ($entity instanceof ProductModelInterface) {
-                if ($this->stepExecution) {
-                    if (!$this->readChildren) {
-                        $warning = 'This bulk action doesn\'t support Product models entities yet.';
-                        $this->stepExecution->addWarning(
-                            $warning,
-                            [],
-                            new DataInvalidItem(['code' => $entity->getCode()])
-                        );
-                    }
-                }
-
-                $entity = null;
-                continue;
-            }
-
             $this->stepExecution->incrementSummaryInfo('read');
-
-            break;
         }
+        $this->firstRead = false;
 
         return $entity;
     }
