+++ src/Akeneo/Pim/Enrichment/Component/spec/Product/Connector/Reader/Database/MassEdit/FilteredProductModelReaderSpec.php
@@ -1,14 +1,12 @@
 <?php
 
-namespace spec\Pim\Bundle\EnrichBundle\Connector\Reader\MassEdit;
+namespace spec\Akeneo\Pim\Enrichment\Component\Product\Connector\Reader\Database\MassEdit;
 
 use Akeneo\Tool\Component\Batch\Job\JobParameters;
 use Akeneo\Tool\Component\Batch\Model\StepExecution;
 use Akeneo\Tool\Component\StorageUtils\Cursor\CursorInterface;
-use Akeneo\Tool\Component\StorageUtils\Detacher\ObjectDetacherInterface;
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\EnrichBundle\Connector\Reader\MassEdit\FilteredProductAndProductModelReader;
-use Akeneo\Pim\Enrichment\Component\Product\Manager\CompletenessManager;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\Reader\Database\MassEdit\FilteredProductModelReader;
 use Akeneo\Channel\Component\Model\ChannelInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductModelInterface;
@@ -19,27 +17,23 @@ use Akeneo\Pim\Enrichment\Component\Product\Converter\MetricConverter;
 use Prophecy\Argument;
 use Prophecy\Promise\ReturnPromise;
 
-class FilteredProductAndProductModelReaderSpec extends ObjectBehavior
+class FilteredProductModelReaderSpec extends ObjectBehavior
 {
     function it_is_initializable()
     {
-        $this->shouldHaveType(FilteredProductAndProductModelReader::class);
+        $this->shouldHaveType(FilteredProductModelReader::class);
     }
 
     function let(
         ProductQueryBuilderFactoryInterface $pqbFactory,
         ChannelRepositoryInterface $channelRepository,
-        CompletenessManager $completenessManager,
         MetricConverter $metricConverter,
         StepExecution $stepExecution
     ) {
         $this->beConstructedWith(
             $pqbFactory,
             $channelRepository,
-            $completenessManager,
-            $metricConverter,
-            true,
-            false
+            $metricConverter
         );
 
         $this->setStepExecution($stepExecution);
@@ -56,16 +50,12 @@ class FilteredProductAndProductModelReaderSpec extends ObjectBehavior
         $channelRepository,
         $metricConverter,
         $stepExecution,
-        $completenessManager,
         ChannelInterface $channel,
         ProductQueryBuilderInterface $pqb,
         CursorInterface $cursor,
         ProductModelInterface $productModel1,
         ProductModelInterface $productModel2,
         ProductModelInterface $productModel3,
-        ProductInterface $product1,
-        ProductInterface $product2,
-        ProductInterface $product3,
         JobParameters $jobParameters
     ) {
         $filters = [
@@ -94,15 +84,18 @@ class FilteredProductAndProductModelReaderSpec extends ObjectBehavior
         $channelRepository->findOneByIdentifier('mobile')->willReturn($channel);
         $channel->getCode()->willReturn('mobile');
 
-        $pqbFactory->create(['filters' => $filters['data'], 'default_scope' => 'mobile'])
+        $pqbFactory->create(['filters' => array_merge($filters['data'], [[
+            'field' => 'entity_type',
+            'operator' => '=',
+            'value' => ProductModelInterface::class,
+        ]]), 'default_scope' => 'mobile'])
             ->shouldBeCalled()
             ->willReturn($pqb);
         $pqb->execute()
             ->shouldBeCalled()
             ->willReturn($cursor);
 
-        $completenessManager->generateMissingForChannel($channel)->shouldBeCalled();
-        $products = [$productModel1, $product1, $productModel2, $product2, $product3, $productModel3];
+        $products = [$productModel1, $productModel2, $productModel3];
         $productsCount = count($products);
         $cursor->valid()->will(
             function () use (&$productsCount) {
@@ -110,21 +103,16 @@ class FilteredProductAndProductModelReaderSpec extends ObjectBehavior
             }
         );
         $cursor->current()->will(new ReturnPromise($products));
-        $cursor->next()->shouldBeCalledTimes(5);
+        $cursor->next()->shouldBeCalled();
 
         $stepExecution->incrementSummaryInfo('read')->shouldBeCalledTimes(3);
         $metricConverter->convert(Argument::any(), $channel)->shouldBeCalledTimes(3);
-
-        $productModel1->getCode()->willReturn('product_model_1');
-        $productModel2->getCode()->willReturn('product_model_2');
-        $productModel3->getCode()->willReturn('product_model_3');
-        $stepExecution->addWarning('This bulk action doesn\'t support Product models entities yet.', Argument::cetera())
-            ->shouldBeCalledTimes(3);
+        $stepExecution->incrementSummaryInfo('skip')->shouldBeCalledTimes(0);
 
         $this->initialize();
-        $this->read()->shouldReturn($product1);
-        $this->read()->shouldReturn($product2);
-        $this->read()->shouldReturn($product3);
+        $this->read()->shouldReturn($productModel1);
+        $this->read()->shouldReturn($productModel2);
+        $this->read()->shouldReturn($productModel3);
         $this->read()->shouldReturn(null);
     }
 }
