+++ src/Akeneo/Pim/Enrichment/Component/spec/Product/Connector/Reader/Database/MassEdit/FilteredProductReaderSpec.php
@@ -1,15 +1,16 @@
 <?php
 
-namespace spec\Pim\Bundle\EnrichBundle\Connector\Reader\MassEdit;
+namespace spec\Akeneo\Pim\Enrichment\Component\Product\Connector\Reader\Database\MassEdit;
 
 use Akeneo\Tool\Component\Batch\Job\JobParameters;
 use Akeneo\Tool\Component\Batch\Model\StepExecution;
 use Akeneo\Tool\Component\StorageUtils\Cursor\CursorInterface;
+use Akeneo\Tool\Component\StorageUtils\Detacher\ObjectDetacherInterface;
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\EnrichBundle\Connector\Reader\MassEdit\FilteredProductModelReader;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\Reader\Database\MassEdit\FilteredProductReader;
+use Akeneo\Pim\Enrichment\Component\Product\Manager\CompletenessManager;
 use Akeneo\Channel\Component\Model\ChannelInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
-use Akeneo\Pim\Enrichment\Component\Product\Model\ProductModelInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Query\ProductQueryBuilderFactoryInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Query\ProductQueryBuilderInterface;
 use Akeneo\Channel\Component\Repository\ChannelRepositoryInterface;
@@ -17,32 +18,30 @@ use Akeneo\Pim\Enrichment\Component\Product\Converter\MetricConverter;
 use Prophecy\Argument;
 use Prophecy\Promise\ReturnPromise;
 
-class FilteredProductModelReaderSpec extends ObjectBehavior
+class FilteredProductReaderSpec extends ObjectBehavior
 {
     function it_is_initializable()
     {
-        $this->shouldHaveType(FilteredProductModelReader::class);
+        $this->shouldHaveType(FilteredProductReader::class);
     }
 
     function let(
         ProductQueryBuilderFactoryInterface $pqbFactory,
         ChannelRepositoryInterface $channelRepository,
+        CompletenessManager $completenessManager,
         MetricConverter $metricConverter,
+        ObjectDetacherInterface $objectDetacher,
         StepExecution $stepExecution
     ) {
         $this->beConstructedWith(
             $pqbFactory,
             $channelRepository,
-            $metricConverter
+            $completenessManager,
+            $metricConverter,
+            true
         );
 
         $this->setStepExecution($stepExecution);
-    }
-
-    function it_set_step_execution(
-        $stepExecution
-    ) {
-        $this->setStepExecution($stepExecution)->shouldReturn(null);
     }
 
     function it_reads_products_only_and_not_product_models(
@@ -50,12 +49,13 @@ class FilteredProductModelReaderSpec extends ObjectBehavior
         $channelRepository,
         $metricConverter,
         $stepExecution,
+        $completenessManager,
         ChannelInterface $channel,
         ProductQueryBuilderInterface $pqb,
         CursorInterface $cursor,
-        ProductModelInterface $productModel1,
-        ProductModelInterface $productModel2,
-        ProductModelInterface $productModel3,
+        ProductInterface $product1,
+        ProductInterface $product2,
+        ProductInterface $product3,
         JobParameters $jobParameters
     ) {
         $filters = [
@@ -72,6 +72,11 @@ class FilteredProductModelReaderSpec extends ObjectBehavior
                         'camcorder',
                     ],
                 ],
+                [
+                    'field' => 'entity_type',
+                    'operator' => '=',
+                    'value' => ProductInterface::class,
+                ],
             ],
             'structure' => [
                 'scope'   => 'mobile',
@@ -87,7 +92,7 @@ class FilteredProductModelReaderSpec extends ObjectBehavior
         $pqbFactory->create(['filters' => array_merge($filters['data'], [[
             'field' => 'entity_type',
             'operator' => '=',
-            'value' => ProductModelInterface::class,
+            'value' => ProductInterface::class,
         ]]), 'default_scope' => 'mobile'])
             ->shouldBeCalled()
             ->willReturn($pqb);
@@ -95,7 +100,8 @@ class FilteredProductModelReaderSpec extends ObjectBehavior
             ->shouldBeCalled()
             ->willReturn($cursor);
 
-        $products = [$productModel1, $productModel2, $productModel3];
+        $completenessManager->generateMissingForChannel($channel)->shouldBeCalled();
+        $products = [$product1, $product2, $product3];
         $productsCount = count($products);
         $cursor->valid()->will(
             function () use (&$productsCount) {
@@ -110,9 +116,9 @@ class FilteredProductModelReaderSpec extends ObjectBehavior
         $stepExecution->incrementSummaryInfo('skip')->shouldBeCalledTimes(0);
 
         $this->initialize();
-        $this->read()->shouldReturn($productModel1);
-        $this->read()->shouldReturn($productModel2);
-        $this->read()->shouldReturn($productModel3);
+        $this->read()->shouldReturn($product1);
+        $this->read()->shouldReturn($product2);
+        $this->read()->shouldReturn($product3);
         $this->read()->shouldReturn(null);
     }
 }
