+++ src/Akeneo/Pim/Enrichment/Component/spec/Product/Connector/Reader/Database/MassEdit/ProductAndProductModelReaderSpec.php
@@ -1,58 +1,61 @@
 <?php
 
-namespace spec\Pim\Bundle\EnrichBundle\Connector\Reader\MassEdit;
+namespace spec\Akeneo\Pim\Enrichment\Component\Product\Connector\Reader\Database\MassEdit;
 
 use Akeneo\Tool\Component\Batch\Job\JobParameters;
 use Akeneo\Tool\Component\Batch\Model\StepExecution;
 use Akeneo\Tool\Component\StorageUtils\Cursor\CursorInterface;
-use Akeneo\Tool\Component\StorageUtils\Detacher\ObjectDetacherInterface;
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\EnrichBundle\Connector\Reader\MassEdit\FilteredProductReader;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\Reader\Database\MassEdit\ProductAndProductModelReader;
 use Akeneo\Pim\Enrichment\Component\Product\Manager\CompletenessManager;
 use Akeneo\Channel\Component\Model\ChannelInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
+use Akeneo\Pim\Enrichment\Component\Product\Model\ProductModelInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Query\ProductQueryBuilderFactoryInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Query\ProductQueryBuilderInterface;
 use Akeneo\Channel\Component\Repository\ChannelRepositoryInterface;
-use Akeneo\Pim\Enrichment\Component\Product\Converter\MetricConverter;
-use Prophecy\Argument;
 use Prophecy\Promise\ReturnPromise;
 
-class FilteredProductReaderSpec extends ObjectBehavior
+class ProductAndProductModelReaderSpec extends ObjectBehavior
 {
     function it_is_initializable()
     {
-        $this->shouldHaveType(FilteredProductReader::class);
+        $this->shouldHaveType(ProductAndProductModelReader::class);
     }
 
     function let(
         ProductQueryBuilderFactoryInterface $pqbFactory,
         ChannelRepositoryInterface $channelRepository,
         CompletenessManager $completenessManager,
-        MetricConverter $metricConverter,
-        ObjectDetacherInterface $objectDetacher,
         StepExecution $stepExecution
     ) {
         $this->beConstructedWith(
             $pqbFactory,
             $channelRepository,
             $completenessManager,
-            $metricConverter,
             true
         );
 
         $this->setStepExecution($stepExecution);
     }
 
-    function it_reads_products_only_and_not_product_models(
+    function it_set_step_execution(
+        $stepExecution
+    ) {
+        $this->setStepExecution($stepExecution)->shouldReturn(null);
+    }
+
+    function it_reads_products_and_product_models(
         $pqbFactory,
         $channelRepository,
-        $metricConverter,
         $stepExecution,
         $completenessManager,
         ChannelInterface $channel,
         ProductQueryBuilderInterface $pqb,
         CursorInterface $cursor,
+        ProductModelInterface $productModel1,
+        ProductModelInterface $productModel2,
+        ProductModelInterface $productModel3,
         ProductInterface $product1,
         ProductInterface $product2,
         ProductInterface $product3,
@@ -72,11 +75,6 @@ class FilteredProductReaderSpec extends ObjectBehavior
                         'camcorder',
                     ],
                 ],
-                [
-                    'field' => 'entity_type',
-                    'operator' => '=',
-                    'value' => ProductInterface::class,
-                ],
             ],
             'structure' => [
                 'scope'   => 'mobile',
@@ -89,11 +87,7 @@ class FilteredProductReaderSpec extends ObjectBehavior
         $channelRepository->findOneByIdentifier('mobile')->willReturn($channel);
         $channel->getCode()->willReturn('mobile');
 
-        $pqbFactory->create(['filters' => array_merge($filters['data'], [[
-            'field' => 'entity_type',
-            'operator' => '=',
-            'value' => ProductInterface::class,
-        ]]), 'default_scope' => 'mobile'])
+        $pqbFactory->create(['filters' => $filters['data'], 'default_scope' => 'mobile'])
             ->shouldBeCalled()
             ->willReturn($pqb);
         $pqb->execute()
@@ -101,7 +95,7 @@ class FilteredProductReaderSpec extends ObjectBehavior
             ->willReturn($cursor);
 
         $completenessManager->generateMissingForChannel($channel)->shouldBeCalled();
-        $products = [$product1, $product2, $product3];
+        $products = [$productModel1, $product1, $productModel2, $product2, $product3, $productModel3];
         $productsCount = count($products);
         $cursor->valid()->will(
             function () use (&$productsCount) {
@@ -109,16 +103,21 @@ class FilteredProductReaderSpec extends ObjectBehavior
             }
         );
         $cursor->current()->will(new ReturnPromise($products));
-        $cursor->next()->shouldBeCalled();
+        $cursor->next()->shouldBeCalledTimes(5);
 
-        $stepExecution->incrementSummaryInfo('read')->shouldBeCalledTimes(3);
-        $metricConverter->convert(Argument::any(), $channel)->shouldBeCalledTimes(3);
-        $stepExecution->incrementSummaryInfo('skip')->shouldBeCalledTimes(0);
+        $stepExecution->incrementSummaryInfo('read')->shouldBeCalledTimes(6);
+
+        $productModel1->getCode()->willReturn('product_model_1');
+        $productModel2->getCode()->willReturn('product_model_2');
+        $productModel3->getCode()->willReturn('product_model_3');
 
         $this->initialize();
+        $this->read()->shouldReturn($productModel1);
         $this->read()->shouldReturn($product1);
+        $this->read()->shouldReturn($productModel2);
         $this->read()->shouldReturn($product2);
         $this->read()->shouldReturn($product3);
+        $this->read()->shouldReturn($productModel3);
         $this->read()->shouldReturn(null);
     }
 }
